refactor: move DTOs into dtos folder
